feat: migrate hooks usage
