+++ src/App.tsx
@@ -1,10 +1,9 @@
-import { Routes, Route, BrowserRouter } from "react-router-dom"
-import HooksLayout from "./HooksLayout"
-import HomePage from "./pages/HomePage"
-import AllHooksPage from "./pages/AllHooksPage"
-import { renderHookRoutes } from "./routes"
-import "./App.css"
+import { BrowserRouter, Route, Routes } from "react-router-dom"
 import useLocalStorage from "use-local-storage"
+import "./App.css"
+import HooksLayout from "./HooksLayout"
+import AllHooksPage from "./pages/AllHooksPage"
+import HomePage from "./pages/HomePage"
 
 function App() {
   const defaultDark = window.matchMedia("(prefers-color-scheme: dark)").matches
@@ -61,7 +60,6 @@ function App() {
           <Route index element={<HomePage />}></Route>
           <Route path="hooks" element={<HooksLayout />}>
             <Route index element={<AllHooksPage />}></Route>
-            {renderHookRoutes}
           </Route>
         </Routes>
       </div>
