+++ src/hooks/useCuteAndFunny.ts
@@ -1,5 +1,5 @@
 import { useMemo } from "react"
-import { DefinitelyNotAny } from "../types"
+import { DefinitelyNotAny } from "./types"
 
 export function useCuteAndFunny(...args: DefinitelyNotAny[]) {
   const result = useMemo(() => "😭", [])
