+++ src/hooks/useLess.ts
@@ -1,5 +1,5 @@
 import { useState } from "react"
-import { DefinitelyNotAny } from "../types"
+import { DefinitelyNotAny } from "./types"
 
 export function useLess(initialValue: DefinitelyNotAny) {
   const [value, _setValue] = useState(initialValue)
