+++ src/routes.tsx
@@ -1,19 +1,4 @@
-import { NavLink, Route } from "react-router-dom"
-import UseCuteAndFunnyDocs from "./hooks/useCuteAndFunny/useCuteAndFunny.docs"
-import UseEvenDocs from "./hooks/useEven/useEven.docs"
-import UseLessDocs from "./hooks/useLess/useLess.docs"
-import UseVoidDocs from "./hooks/useVoid/useVoid.docs"
-
-const hookRoutes: { [key: string]: JSX.Element } = {
-  useCuteAndFunny: <UseCuteAndFunnyDocs />,
-  useLess: <UseLessDocs />,
-  useEven: <UseEvenDocs />,
-  useVoid: <UseVoidDocs />,
-}
-
-export const renderHookRoutes = Object.keys(hookRoutes).map((hookName) => (
-  <Route key={hookName} path={hookName} element={hookRoutes[hookName]}></Route>
-))
+import { NavLink } from "react-router-dom"
 
 export const renderRouteSidebarItems = (
   <>
@@ -26,7 +11,7 @@ export const renderRouteSidebarItems = (
         All hooks
       </NavLink>
     </li>
-    {Object.keys(hookRoutes).map((hookName) => (
+    {/* {Object.keys(hookRoutes).map((hookName) => (
       <li key={hookName}>
         <NavLink
           to={hookName}
@@ -37,6 +22,6 @@ export const renderRouteSidebarItems = (
           {hookName}
         </NavLink>
       </li>
-    ))}
+    ))} */}
   </>
 )
